+++ riscv/src/register.rs
@@ -106,6 +106,14 @@ pub mod minstreth;
 mod mhpmeventx;
 pub use self::mhpmeventx::*;
 
+// Pick some registers based on architecture
+#[cfg(feature = "ballast-sysctrl")]
+mod ballast_sysctrl;
+#[cfg(feature = "ballast-mpc")]
+mod ballast_mpc;
+#[cfg(feature = "tackle")]
+mod tackle;
+
 // TODO: Debug/Trace Registers (shared with Debug Mode)
 
 // TODO: Debug Mode Registers